requirements.txt
python-telegram-bot==13.15
qrcode[pil]
fastapi
uvicorn
SQLAlchemy
Pillow
aiofiles
